Use JSON.generate in favor of to_json
to_json is from rails which might not be around. It just happens to be here because we explicitly test against rails versions.
