+++ spec/flipper/api/json_params_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Flipper::Api::JsonParams do
   let(:app) do
     app = lambda do |env|
       request = Rack::Request.new(env)
-      [200, { 'Content-Type' => 'application/json' }, [request.params.to_json]]
+      [200, { 'Content-Type' => 'application/json' }, [JSON.generate(request.params)]]
     end
     builder = Rack::Builder.new
     builder.use described_class
@@ -15,7 +15,7 @@ RSpec.describe Flipper::Api::JsonParams do
   describe 'json post request' do
     it 'adds request body to params' do
       response = post '/',
-                      { flipper_id: 'user:2' }.to_json,
+                      JSON.generate({ flipper_id: 'user:2' }),
                       'CONTENT_TYPE' => 'application/json'
 
       params = JSON.parse(response.body)
@@ -24,7 +24,7 @@ RSpec.describe Flipper::Api::JsonParams do
 
     it 'handles request bodies with multiple params' do
       response = post '/',
-                      { flipper_id: 'user:2', language: 'ruby' }.to_json,
+                      JSON.generate({ flipper_id: 'user:2', language: 'ruby' }),
                       'CONTENT_TYPE' => 'application/json'
 
       params = JSON.parse(response.body)
@@ -33,7 +33,7 @@ RSpec.describe Flipper::Api::JsonParams do
 
     it 'handles request bodies and single query string params' do
       response = post '/?language=ruby',
-                      { flipper_id: 'user:2' }.to_json,
+                      JSON.generate({ flipper_id: 'user:2' }),
                       'CONTENT_TYPE' => 'application/json'
 
       params = JSON.parse(response.body)
@@ -42,7 +42,7 @@ RSpec.describe Flipper::Api::JsonParams do
 
     it 'handles request bodies and multiple query string params' do
       response = post '/?language=ruby&framework=rails',
-                      { flipper_id: 'user:2' }.to_json,
+                      JSON.generate({ flipper_id: 'user:2' }),
                       'CONTENT_TYPE' => 'application/json'
 
       params = JSON.parse(response.body)
@@ -51,7 +51,7 @@ RSpec.describe Flipper::Api::JsonParams do
 
     it 'favors request body params' do
       response = post '/?language=javascript',
-                      { flipper_id: 'user:2', language: 'ruby' }.to_json,
+                      JSON.generate({ flipper_id: 'user:2', language: 'ruby' }),
                       'CONTENT_TYPE' => 'application/json'
 
       params = JSON.parse(response.body)
