+++ spec/flipper/api/v1/actions/percentage_of_time_gate_spec.rb
@@ -39,7 +39,7 @@ RSpec.describe Flipper::Api::V1::Actions::PercentageOfTimeGate do
     before do
       flipper[:my_feature].enable_percentage_of_time(10)
       delete '/features/my_feature/percentage_of_time',
-             { percentage: '5' }.to_json,
+             JSON.generate({ percentage: '5' }),
              'CONTENT_TYPE' => 'application/json'
     end
